+++ Localization/zhCN.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-12-18 17-52-31 ]] --
+-- [[ Exported at 2022-12-20 09-48-30 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --
 
@@ -441,4 +441,10 @@ L["Siege on Dragonbane Keep"] = "围攻灭龙要塞";
 L["Community Feast"] = "社区盛宴";
 L["Needs to be petted for achievement"] = "{forAchievement} 需要抚摸";
 L["Already petted for achievement"] = "{forAchievement} 已抚摸";
+L["%d/%m/%Y %I:%M %p"] = "%d/%m/%Y %I:%M %p";
+L["%m/%d/%Y %R"] = "%m/%d/%Y %R";
+L["%m/%d/%Y %I:%M %p"] = "%m/%d/%Y %I:%M %p";
+L["%Y/%m/%d %R"] = "%Y/%m/%d %R";
+L["%Y/%m/%d %I:%M %p"] = "%Y/%m/%d %I:%M %p";
+L["%c"] = "%c";