+++ Localization/zhCN.ExtraWrathStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-12-20 09-48-30 ]] --
+-- [[ Exported at 2022-12-20 14-56-57 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --